pyproject.toml
[project]
name = "selflearning-platform"
version = "0.1.0"
description = "Платформа для самообучения на Django и DRF с использованием современных технологий"
authors = [
    {name = "Your Name", email = "your.email@example.com"}
]
license = "MIT"
readme = "README.md"
requires-python = ">=3.13"
keywords = ["education", "django", "drf", "postgresql", "jwt"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Web Environment",
    "Framework :: Django",
    "Framework :: Django :: 5.2",
    "Intended Audience :: Education",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3.13",
    "Topic :: Education",
]

dependencies = [
    "django>=5.2.1,<6.0.0",
    "python-dotenv>=1.1.0,<2.0.0",
    "psycopg2-binary>=2.9.10,<3.0.0",
    "djangorestframework>=3.16.0,<4.0.0",
    "djangorestframework-simplejwt>=5.5.0,<6.0.0",
    "drf-yasg>=1.21.10,<2.0.0",
    "django-cors-headers>=4.7.0,<5.0.0",
    "django-filter>=25.1,<26.0",
    "pytest>=8.0.0,<9.0.0",
    "pytest-django>=4.8.0,<5.0.0",
    "pytest-cov>=4.1.0,<5.0.0",
    "black>=24.1.0,<25.0.0",
    "flake8>=7.2.0,<8.0.0",
    "isort>=5.13.0,<6.0.0",
    "mypy>=1.8.0,<2.0.0"
]

[project.urls]
Homepage = "https://eduflow.dev"
Documentation = "https://docs.eduflow.dev"
Repository = "https://github.com/eduflow/eduflow"
"Bug Tracker" = "https://github.com/eduflow/eduflow/issues"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88
target-version = ["py313"]
include = '\.pyi?$'

[tool.isort]
profile = "black"
multi_line_output = 3
line_length = 88

[tool.mypy]
python_version = "3.13"
strict = true
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "config.settings"
python_files = ["test_*.py", "*_test.py"]
addopts = "--cov=. --cov-report=term-missing"
testpaths = ["tests"]
